+++ src/main/java/me/aiglez/lonkskit/players/impl/MemoryLocalPlayer.java
@@ -231,7 +231,7 @@ public class MemoryLocalPlayer implements LocalPlayer {
     }
 
     @Override
-    public double getPoints() {
+    public int getPoints() {
         return this.offline.getPoints();
     }
